+++ infra/output.tf
@@ -2,12 +2,12 @@
 # Outputs -------------------------------------------------------------------------------------
 # ---------------------------------------------------------------------------------------------
 
-output "latency_app_public_ip" {
-  value       = google_compute_instance.latency_app.network_interface[0].access_config[0].nat_ip
-  description = "Public IP of the latency application VM"
+output "probe_public_ip" {
+  value       = google_compute_instance.alien_probe.network_interface[0].access_config[0].nat_ip
+  description = "Public IP of the probe VM"
 }
 
-output "latency_target_private_ip" {
-  value       = google_compute_instance.latency_target.network_interface[0].network_ip
-  description = "Private IP of the latency target VM"
+output "target_private_ip" {
+  value       = google_compute_instance.alien_target.network_interface[0].network_ip
+  description = "Private IP of the target VM"
 }