+++ CHANGELOG.md
@@ -6,10 +6,10 @@ Bump project version into `0.5.1`
 
 - [x] (chore) Project layout components
 
-- [x] (application) Build application logic within development
+- [x] (infrastructure) Provision infrastructure for project within GCP
 
 - [x] (docs) Updated the documentation for project (README, sad, changelog, etc.)
 
-- [x] Build application of probe (agent, collector) instance
+- [x] (application) Build applications of probe instance, included: `agent`, `collector`
 
-- [x] Build application for target (for HTTP call) with single `/_/health` endpoint
+- [x] (application) Build application for target instance, included: `exposer` with single `/_/health` endpoint